requirements/dev.txt
black==22.3.0
docker==5.0.3
psutil==5.8.0
python-dotenv==1.0.0
requests==2.25.1
rich==13.4.2
click==8.0.4
typer[all]==0.9.0
torch==2.0.1
transformers==4.31.0
langchain==0.0.238
einops==0.6.1
accelerate==0.21.0
xformers==0.0.20
safetensors==0.3.1
fastapi==0.100.0
pydantic==1.10.11
uvicorn==0.23.1
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.8.0
celery[redis]==5.2.2
streamlit==1.25.0
streamlit-extras==0.2.7
protobuf==4.23.4
